// src/types.ts
// Типы данных для приложения Kanban

// Возможные статусы задачи (колонки доски)
export type Status = 'todo' | 'wip' | 'testing' | 'done';

// Описание сущности задачи
export interface Task {
  // Уникальный идентификатор. Нужен для DnD и обновлений
  id: string;
  // Заголовок/название задачи, вводится пользователем
  title: string;
  // Дата создания в ISO-формате (строка), заполняется автоматически
  createdAt: string;
  // новое поле  приоритет priority
  priority: 'low' | 'medium' | 'high';
  // Статус задачи (в какой колонке она находится сейчас)
  status: Status;
}